Only run background_work on an HPX thread and limit number of sends active
Mixing locks for OS threads and hpx threads caused too much trouble as
suspending an hpx thread is ok, but suspending an OS thread can cause problems.

Limit the number of active sends permitted to avoid exhausing the memory pool
or the rdma send queues. The limit needs to be assessed to find the best
trade off between waits and memory consumption.